Add Config Options
TSC now checks for unused local variables, implicit returns in a function, and fallthrough cases in a switch statement. The TSC also removes comments from the source. Even though allowJS is false by default, this was added for the sake of transparency.

The package.json now defines the repository and a start script for the project.